Created Background class. More info in extended description.
The Background class extends Entity class, and is used for making
the scrolling backgrounds around Red. One Background class object
will need to be created per background layer. This class may or may
not be editted in the future, but as of right now it will require
a .png image that is 1024-pixels wide by y-pixels tall (it can be any height).

I can't do it now, I'm in a pretty big hurry, but I will go back through
and comment out the Background class when I next have some time.